LayoutTests/imported/w3c/web-platform-tests/css/css-ruby/br-clear-all-001-expected.txt

The yellow box should encompass its contents and also the cyan float, due to clearance. The yellow box has bottom padding, and the blue ruby annotation box is allowed to overflow into the padding area.



FAIL #container 1 assert_equals:
<div id="container" data-expected-height="150">
  <ruby>
    <div></div>
    <rt>
      <div></div>
    </rt>
  </ruby>
  <br clear="all">
</div>
height expected 150 but got 171
